# openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "API Documentation",
    "description" : "API với xác thực Bearer Token",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/profile-documents/{id}" : {
      "put" : {
        "tags" : [ "profile-document-controller" ],
        "operationId" : "updateDocument",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "name",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "note",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "removeFiles",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "newFiles" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string",
                      "format" : "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileDocument"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "profile-document-controller" ],
        "operationId" : "deleteDocument",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/profile-documents" : {
      "post" : {
        "tags" : [ "profile-document-controller" ],
        "operationId" : "createDocument",
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "note",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "profileId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "files" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string",
                      "format" : "binary"
                    }
                  }
                },
                "required" : [ "files" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileDocument"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/profiles/{id}" : {
      "get" : {
        "tags" : [ "profile-controller" ],
        "operationId" : "getProfileById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/profile-documents/{profileId}" : {
      "get" : {
        "tags" : [ "profile-document-controller" ],
        "operationId" : "getDocuments",
        "parameters" : [ {
          "name" : "profileId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProfileDocumentDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/medical-records/profile/{profileId}" : {
      "get" : {
        "tags" : [ "medical-record-controller" ],
        "operationId" : "getAllMedicalRecords",
        "parameters" : [ {
          "name" : "profileId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/MedicalRecordDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/files/{id}/{filename}" : {
      "delete" : {
        "tags" : [ "file-controller" ],
        "operationId" : "deleteFile",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "filename",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ProfileDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "bloodType" : {
            "type" : "string"
          },
          "height" : {
            "type" : "number",
            "format" : "double"
          },
          "weight" : {
            "type" : "number",
            "format" : "double"
          },
          "allergies" : {
            "type" : "string"
          },
          "chronicDiseases" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "healthInsuranceNumber" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "phoneNumber" : {
            "type" : "string"
          },
          "dateOfBirth" : {
            "type" : "string",
            "format" : "date"
          },
          "gender" : {
            "type" : "string",
            "enum" : [ "FEMALE", "MALE" ]
          },
          "email" : {
            "type" : "string"
          },
          "address" : {
            "type" : "string"
          },
          "avatarUrl" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "Family" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "familyName" : {
            "type" : "string",
            "maxLength" : 255,
            "minLength" : 3
          },
          "address" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "owner" : {
            "$ref" : "#/components/schemas/Profile"
          },
          "ownerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "phoneNumber" : {
            "type" : "string",
            "pattern" : "^(\\+84|0)[3|5|7|8|9][0-9]{8}$"
          },
          "email" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "familyMembers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FamilyMember"
            }
          }
        },
        "required" : [ "address", "familyName", "isActive", "owner" ]
      },
      "FamilyMember" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "family" : {
            "$ref" : "#/components/schemas/Family"
          },
          "familyId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "profile" : {
            "$ref" : "#/components/schemas/Profile"
          },
          "profileId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "relationship" : {
            "type" : "string",
            "maxLength" : 50,
            "minLength" : 0
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "isActive" : {
            "type" : "boolean"
          }
        },
        "required" : [ "family", "relationship" ]
      },
      "GrantedAuthority" : {
        "type" : "object",
        "properties" : {
          "authority" : {
            "type" : "string"
          }
        }
      },
      "Profile" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "bloodType" : {
            "type" : "string",
            "pattern" : "^(A|B|AB|O)[+-]$"
          },
          "height" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 300.0,
            "minimum" : 0.0
          },
          "weight" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 500.0,
            "minimum" : 0.0
          },
          "allergies" : {
            "type" : "string",
            "maxLength" : 1000,
            "minLength" : 0
          },
          "chronicDiseases" : {
            "type" : "string",
            "maxLength" : 1000,
            "minLength" : 0
          },
          "notes" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          },
          "healthInsuranceNumber" : {
            "type" : "string",
            "maxLength" : 20,
            "minLength" : 0
          },
          "fullName" : {
            "type" : "string"
          },
          "phoneNumber" : {
            "type" : "string"
          },
          "dateOfBirth" : {
            "type" : "string",
            "format" : "date"
          },
          "gender" : {
            "type" : "string",
            "enum" : [ "FEMALE", "MALE" ]
          },
          "email" : {
            "type" : "string"
          },
          "address" : {
            "type" : "string"
          },
          "avatarUrl" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/User"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "familyMembers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FamilyMember"
            }
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "attachments" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "ProfileDocument" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          },
          "attachments" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "profile" : {
            "$ref" : "#/components/schemas/Profile"
          },
          "profileId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "User" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "password" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "lastLogin" : {
            "type" : "string",
            "format" : "date-time"
          },
          "profile" : {
            "$ref" : "#/components/schemas/Profile"
          },
          "username" : {
            "type" : "string"
          },
          "authorities" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GrantedAuthority"
            }
          },
          "enabled" : {
            "type" : "boolean"
          },
          "credentialsNonExpired" : {
            "type" : "boolean"
          },
          "accountNonExpired" : {
            "type" : "boolean"
          },
          "accountNonLocked" : {
            "type" : "boolean"
          }
        }
      },
...
      "DoctorDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "specialty" : {
            "type" : "string"
          },
          "licenseNumber" : {
            "type" : "string"
          },
          "medicalFacility" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          },
          "isVerified" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "MedicalRecordDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "profile" : {
            "$ref" : "#/components/schemas/ProfileDTO"
          },
          "title" : {
            "type" : "string"
          },
          "profileId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "visitDate" : {
            "type" : "string",
            "format" : "date"
          },
          "medicalFacility" : {
            "type" : "string"
          },
          "doctorName" : {
            "type" : "string"
          },
          "doctor" : {
            "$ref" : "#/components/schemas/DoctorDTO"
          },
          "diagnosis" : {
            "type" : "string"
          },
          "treatment" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "isFollowup" : {
            "type" : "boolean"
          },
          "followupDate" : {
            "type" : "string",
            "format" : "date"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createdBy" : {
            "type" : "string"
          },
          "attachments" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "ProfileDocumentDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          },
          "attachments" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "profileId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}